fix: address pr ci failures
