+++ apps/desktop/src/api/clientWorkflowGraph.ts
@@ -26,10 +26,7 @@ export function workflowGraphDraftPayload(graph: WorkflowGraphDraft): JsonObject
           node.completionMode !== undefined && node.completionMode.length > 0
             ? node.completionMode
             : undefined,
-        script_path:
-          typeof node.scriptPath === "string" && node.scriptPath.trim().length > 0
-            ? node.scriptPath
-            : undefined,
+        script_path: node.scriptPath?.trim().length ? node.scriptPath : undefined,
         input_fields: node.inputFields.map((field) => ({
           name: field.name,
           description: field.description,
